work of 25.8

+++ maximize_modularity.h
@@ -19,7 +19,7 @@
 
 double calc_modularity_delta(leading_eigenpair* eigenpair);
 double improve_modularity(leading_eigenpair* eigenpair);
-void find_maximal_score(max_in_array *max_score, int group_size, int* unmoved, leading_eigenpair* eigenpair, int q0);
+void find_maximal_score(max_in_array *max_score, int group_size, int* unmoved, leading_eigenpair* eigenpair, double q0);
 double calc_final_division(int group_size, max_in_array *max_improvemnt, double *division_vector, int *indices);
